lintian: Back out of loading checks from $HOME and /etc
aptdaemon uses Lintian to do a minimal check of packages before
installing them.  It is possible that loading checks from $HOME and
/etc by default could lead to (local user) privilege escalation.

The upstream of aptdaemon has been notified of the coming changes[1],
but they might not be the only ones to use Lintian in a situation like
this.  So only allow "non-code" files to be loaded from $HOME and /etc
by default for now.  Where needed/desired, it is still possible to use
--include-dir to "promote" these directories.

[1] https://bugs.launchpad.net/aptdaemon/+bug/1162947

Signed-off-by: Niels Thykier <niels@thykier.net>