L::Util::touch_file: Pass the filehandle to utime
Pass the open filehandle to utime instead of the file name.  This
ensures that we change the utime of the "correct" file (i.e. we avoid
a possible race condition).

The race condition could not occur with Lintian current use of this
sub, but that might change and others might use the code as reference.

Signed-off-by: Niels Thykier <niels@thykier.net>