L::Collect::*: Allow L::Path objects as arg to unpacked/control
In unpacked and control allow callers to pass L::Path object instead a
string.  It allows these methods to skip some sanity checking of the
input.

At the same time forbit an explicit "undef" as argument.  Previously,
it would not have been a problem, but now it could hide a bug in the
code (namely, a failure to check the return value of "index").

Signed-off-by: Niels Thykier <niels@thykier.net>