c/cruft: Don't consider all ../../** symlinks unsafe
$_ in find_cruft is set to the basename of the examined file.  The
effect of the link destination on basename($_) is currently used in
order to check whether a symlink escapes the root directory of the
source package; this results in false positives for all symlinks moving
up more than one level, as basename($_) is always ".".  Base this check
on the full relative path to the file ($name) instead.

[nthykier: Fixup the test with an incorrect tags file]

Signed-off-by: Niels Thykier <niels@thykier.net>