Improve the version information of packages bundled with the Perl core
There are various incompatibilities with the CPAN and Debian versioning
schemes. Try to be smarter about deriving Debian version numbers from
Module::CoreList data, taking epochs and varying number of decimal digits
into account.

Most of this code is reused from the Debian maintainer tests of the
Perl package (debian/t/control.t in the 'perl' source package, 5.14.2-21
or so at the time of this writing.)

Signed-off-by: Niels Thykier <niels@thykier.net>