Fix 'package-superseded-by-perl' epoch handling
The perl-provides data now includes epochs, so we have
to take those into account when comparing the package version
to the one in the Perl core.

The test relies on the fact that libversion-perl currently has an epoch
in the Debian archive. The old code would strip the test package epoch
before the comparison, so 9:999 was falsely smaller than 1:0.8800.

Signed-off-by: Niels Thykier <niels@thykier.net>