L::Util: Give detailed line information from dctrl files
Change parse_dpkg_control and visit_dpkg_control to use a hashref
instead of a single integer for line number information.  The hashref
records the line number of each field individually as well as the line
number of the first field (via the "START-OF-PARAGRAPH" key).

Updated c/source-copyright.pm to cope with this change and use more
precise line numbers where possible.

Signed-off-by: Niels Thykier <niels@thykier.net>