t: Add optional "Test-Depends" field
Add support for the (optional) field, "Test-Depends".  If present, the
dependencies listed in the field will be checked.  If the dependency
field is unsatisfiable, the test is skipped.

This is useful for skipping tests that cannot be done in (e.g.)
stable, where some packages are missing or too old.

Signed-off-by: Niels Thykier <niels@thykier.net>