checks/cruft.pm: Ignore README.Debian file
Usually README.Debian does not contains an explicit license, it uses
the package license defined in debian/copyright. So, we can avoiding
the "json-evil" check (poor example) and use the following text
without declaring an override:

"The package originally comes with XXX and this file is licensed under
MIT license with one added sentence: 'The Software shall be used for Good,
not Evil.' However, it violates the Debian Policy."

If "changelog are changelog", README.Debian are README.Debian.

Signed-off-by: Niels Thykier <niels@thykier.net>