c/cruft.pm: Fix a false positive in gtk-doc
This is a really hard case this time. The example is like this:
> GNU Free Documentation license
> Sweedish text
> GNU Free Documentation license
> english text of invariants
> a copy is included.

In order to get it we need to match the interesting section
($gfdlsections in the code or here english text of invariant) only if
this section does not match GNU Free Documentation license. So we need
to change

  .*? by (?:(?!gnu \s+ free \s+ documentation \s+ license).)*?.

Unfortunalty it means regression on one of the text case maintain.texi
that is an example text.  In order to fix this regression we need to
get more context arround the GNU Free Documentation license pattern.

So if the context before means example do not tag.

Last but not least they are a corner case. Because we do a sliding
windows match, we could loose the context before in the block recently
discarded. Thus except if it is the first block, always ask for 1024
char exactly of context.

[nthykier: replace anonymous sub with a named one]

Signed-off-by: Niels Thykier <niels@thykier.net>