Move all legacy tests to t/tests/legacy-*
With this move, all legacy tests are now run during the "suite:tests"
run of the regular tests.  They all have the sequence "9999", so they
are still the last tests to be run.

Note that this means that "onlyrun=--legacy-only" is replaced by
"onlyrun=legacy".

This change means that:
 * test failures of legacy tests now causes FTBFS
 * legacy tests are now run in parallel
 * fewer test-runners are needed \o/

The legacy tests are still counted separated when it comes to
"test^Wtag coverage" of the test suite.

Signed-off-by: Niels Thykier <niels@thykier.net>