L::Relation: pkg now implies pkg:any (but not vice versa)
When the "q"-relation has an ":any" multi-arch qualifier, allow the
"p"-relation to still imply the "q"-relation as if "q" did not have
":any" qualification.  However, the case where the "p"-relation has
an ":any" qualification remains unchanged.

The rationale behind this change is that a dependency on "python" is
(AFAICT) stronger than "python:any".  The minor difference between
that "python:any" makes no restrictions on the architecture of the
dependency whereas "python" requires the dependency to have same
architecture as the package depending on it.

Signed-off-by: Niels Thykier <niels@thykier.net>