L::C::Binary: Add support for multi-component RPATHs
The RPATH setting works similar to PATH, and supports multiple
components, separated by a colon. As such, Lintian should be able to
handle those, and split the RPATH into components before making any
checks on those.

This patch does just that, and adds a test case to trigger the original
issue too.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>