Fix link rel false positive
When run on the remctl source package, Lintian reports:

X: remctl-client: privacy-breach-generic usr/share/doc/remctl-client/protocol.html purl.org/dc/terms/

The relevant line of the documentation is:

 <link rel="schema.dct" href="http://purl.org/dc/terms/" />

This is used to provide a canonical schema for <meta> element naming
using Dublin Core.  This is akin to the reference to w3c.org in the
SGML declarations: it's a URI to establish a namespace, but is not a
URL loaded by a browser while displaying the page, and therefore is not
a privacy breach.

Closes: #736992
  <link rel="schema.dct" href="http://purl.org/dc/terms/" />