r/html_reports: Optimise away some regex subst
Use index() to skip some s/// invocations.  When profiling
html_reports, this seems to reduce the runtime with about 30s.

For reference, html_quote is called about 5.6 million times, but the
"escapble" characters are only found in 9-10k, ~7k and ~7k of strings
checked (respectively).  In other words, no more than 0.5% of the
strings need any escaping (in practise).

Signed-off-by: Niels Thykier <niels@thykier.net>