L::Util: Use hack to optimise {,r}strip
It is "well-known" that the expression s/^\s++|\s++$/g is slower than
splitting the expression into to separate ones (namely, s/^\s++// and
s/\s++$//)[1].

During profiling, it turns out that the s/\s++$// part is the slowest
part.  This can "fortunately" be replaced by the completely unreadable
alternative "unpack('A*', $string)", which sadly appears to be at least
a factor 10 better than s/\s++$//.

Kudos to Ansgar Burchardt for the suggestion of using unpack.

[1] http://www.perlmonks.org/?node_id=865344

Signed-off-by: Niels Thykier <niels@thykier.net>