runtests: Cache test artifacts in the "tests" suite
After compiling a package built from a "tests" testcase, add a
"build-stamp"-file.  When (re-)running the test, check for this stamp
file and use it to determine whether it makes sense to recompile the
package.

To avoid "funny results" now that the test dir is no longer
unconditionally purged, all artifacts from the suite known as "tests"
is now moved into its own subfolder in "RUNDIR".

Signed-off-by: Niels Thykier <niels@thykier.net>