c/deb-format.pm: Warn on deprecated compression formats for data.tar
The lzma and bzip2 compressions are deprecated, and dpkg-deb emits
warnings (since 1.16.4 and 1.17.7 respectively) when trying to build
such packages.

We emit the new tag uses-deprecated-compression-for-data-tarball for
these, instead of reusing malformed-deb-archive, because these packages
are just fine and will keep being supported at extraction time for the
foreseeable future, even when creating them is disallowed.

Signed-off-by: Guillem Jover <guillem@debian.org>
Signed-off-by: Niels Thykier <niels@thykier.net>