Avoid warning for bookmark link
in fftw3 we get warnings for these:

<LINK rel="Bookmark" title="FFTW Home Page"
      href="http://www.fftw.org">
<LINK rel="Bookmark" title="FFTW Manual"
      href="http://www.fftw.org/doc/">

The worst that probably can happen here is that a browser bookmarks the
upstream url instead of the local help page.
I think this is a minor breach hardly worth the trouble patching. Also
firefox and opera both ignore the link and bookmark the local page, the
spec is not very clear on what a browser should do with this tag.

Signed-off-by: Bastien ROUCARIÈS <roucaries.bastien+debian@gmail.com>