c/fields.pm: Fix a bug with versioned obsolete dependencies
The code for checking if an obsolete dependency had a subtle bug,
where it could downgrade an "(build-)depends-on-obsolete-package"
to an "ored-(build-)depends-on-obsolete-package".

The problem occured in condition of the following "if"-statement:

  if (   $pkg eq $alternatives[0]->[0]
         or $all_obsolete) {
         tag 'build-depends-on-obsolete-package',
           "$field: $pkg${replacement}";
  } else {
         tag 'ored-build-depends-on-obsolete-package',
           "$field: $pkg${replacement}";
  }

Here "$pkg" would be the full "clause" with version.  It was compared
with $alternatives[0]->[0], which is always just the package name.

The bug only appears if there are non-obsolete alternatives, since if
all of the alternatives are obsolete, the "else"-path cannot be taken.
According to trigger this bug, you would need a relation like:

  Build-Depends: x-dev (>= 1.0) | non-obsolete-package

Signed-off-by: Niels Thykier <niels@thykier.net>