Detect undefined license
Some cpyright file references a license that is not defined.
Due to copyright law or treaties, files that are not available under
a defined license are non-free and non-re-distributable.

Referencing an undefined license could mean the file cannot be
distributed in Debian or it could simply mean that the existing
license needs to be documented. In both cases,
the copyright file should be updated to reflect reality.

Signed-off-by: Bastien ROUCARIÈS <roucaries.bastien+debian@gmail.com>