Simplify and merge some "my" declarations
Remove explicit assignments on declaration of "my" variables, when
Perl would normally have initialise them to the same value in its
absence.  This makes Perl produce a slightly simplier OP-code graph[1],
but also allows us/Perl to merge declarations.  Finally, it makes our
declarations consistent across the entire codebase.

[1] This may seem weird at first, but Perl insists on keeping the
assignment. The assignment lead to semantically different behaviour
when people use "goto label;".  This is a practise not found in the
Lintian codebase, so we can safely remove the assignments.

Signed-off-by: Niels Thykier <niels@thykier.net>