L::P::FSInfo: New file for keeping track of common L::Path info
The new class Lintian::Path::FSInfo keeps track of common information
for Lintian::Path objects, such as where the underlying file system
store is.

It also maintains the knowledge for whether absolute paths are
well-defined or not.  Currently, this is always true.  However, it
should be false for e.g. L::Path objects returned from control_index.
This correction will come in a future commit.

Signed-off-by: Niels Thykier <niels@thykier.net>