L::Path: Change fs_path requirements and retire is_valid_path
The fs_path method asserts "_check_open" for anything but directories.
Here it falls back to "_check_access".  The "is_valid_path" method has
been retired as it is now equal to "defined($path->resolve_path)".

The reason for the change is that direct access to files usually leads
to them being opened.

Signed-off-by: Niels Thykier <niels@thykier.net>