clean_env: Stop perserving LOCPATH and INTLTOOL_EXTRACT
We no longer rely on these environment variables (or when we do, we
set them directly in the child needing them).

Note that LOCPATH was added in commit 9b7c089 back when we needed a
"private" en_US.UTF-8 locale.  Since Debian/Wheezy, we no longer need
this, since libc provides us with C.UTF-8.

Signed-off-by: Niels Thykier <niels@thykier.net>