lintian: Add --perf-debug and --perf-output options
Which can be used to log performance related data and writing it to a
separate file (or file descriptor).

This commit also changes the debug behaviour to keep the code simple.
In particular, runtime statistics are now shown already with a single
"--debug" argument and some more performance data is emitted to the
debug log as well.

Signed-off-by: Niels Thykier <niels@thykier.net>