lintian: Add new "developer" options group and optimise --help
Add a new "developer/special usage options" group in the --help output
and in the manpage.  This group and the "Package selection options"
are not shown by default with --help.  Instead this now requires a
"--help=extended" (which --help will tell people).

This reduces the default "--help" output to "just" 44 lines.

Signed-off-by: Niels Thykier <niels@thykier.net>