c/r-lintian-harness: Sort work_splice before printing it
The reporting-lintian-harness command sorts its worklist by
"out-of-dateness" and passes a splice of that to lintian.  Lintian
sorts its worklist and processes it in named order.

This lead to the following (fairly confusing) output:

"""
Running Lintian (round 1/2) ...
  - Range: GROUP:afl/1.44b-1 ... GROUP:dispcalgui/2.5.0.0-1
[...]
   [lintian] processed wildmidi/0.3.8-1 successfully (time: 51.572s)
"""

This is resolved by sorting the work splice before printing it to the
log file.

Signed-off-by: Niels Thykier <niels@thykier.net>