L::Util: Use YAML::XS instead of YAML::Any (#784639)
The idea behind YAML::Any is great in theory.  In practise, the
interface is underspecified and includes a (for us) deal breaking
flaw.

YAML::Any does not specify any particular requirements for loading
YAML with UTF-8 encoding.  This has lead different implementations to
solve this differently.  On one side (YAML::XS), you are supposed to
give a "raw byte stream" and it will fail loudly if you give it an
UTF-8 character stream.  On the other side, you are expected to give
it an UTF-8 character stream if you want UTF-8.

Signed-off-by: Niels Thykier <niels@thykier.net>