c/manpages: Trivial batch check of lexgrog and man
The lexgrog and man page commands seems to primarily have a slow start
up time.  We can work around that by scheduling a number of these
processes, which reduces the serial waiting time a bit.

I chose two small prime numbers to determine how many proceses each
can have running before we start reaping them.  This should minimize
the number of times we wait for *both* processes in the same loop
interation.  Currently, increasing the number of these primes do not
seem to affect the performance significantly.

For the "future work"-department: This batch thing could be smarter.
E.g. we could do a select(2) on the handles and only process the
children that are ready to talk to us.

Signed-off-by: Niels Thykier <niels@thykier.net>