binaries: limit LFS symbol checks to unresolved symbols
Guillem reported in BTS [0] that inetutils-ftpd is marked as
binary-file-built-without-LFS-support while it has LFS support. The
false-positive comes from a local symbol.
This patch limits the symbol checks to symbols which are unresolved
(i.e. not local, come from an external library) as suggested by Guillem.
Another (additional) limitation could be to limit the version to ^GLIBC_
as all the symbols for the LFS check come from GLIBC.

With this patch the inetutils-ftpd package is no longer marked with
binary-file-built-without-LFS-support while the havp package is still
marked.

[0] https://bugs.debian.org/787853
Signed-off-by: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>
Signed-off-by: Niels Thykier <niels@thykier.net>