Fix f/p in perl-module-name-not-mentioned-in-description (Closes: #795779)
index() was called with its parameter order being the wrong way. This
was not present in pkg-perl-tools as using index() replaced the
previously used regular expression matching for performance reasons.

This also adds a test case for such a false positive, i.e. there's now
a package in the test suite which shouldn't trigger the tag.