fix common lic. false pos. in new-style copyright
Fix false positives for lack of common license references in new-style
copyright files when a license refers to another license, by trying to
parse the file and then checking both the names of the licenses and the
texts.

Add new test for references to common licenses as well.

The test suite passes with these changes.

Adds a new dependency on the Data::Alias Perl module, to avoid copying
large texts in memory.

Signed-off-by: Niels Thykier <niels@thykier.net>