data/fields/binary-fields: Build-Ids is a known field
Paul Richard I by the Grace of God of the United Kingdom of Debian and
Ubuntu and of his other realms and territories King Head of the
Fluxbox Window Manager Defender of the Faith Tagliamonte graced
debhelper with His Build-Ids field.  It would be down right rude to
imply that the work of His Greatness is unknown.

Signed-off-by: Niels Thykier <niels@thykier.net>