L::C::B: Reduce the memory usage for objdump_info
A lot of the fields in objdump_info are initialized to a default
value.  Sadly, this has a considerable overhead.  With this patch, we
do away with most defaults.  This reduces the memory usage of
objdump_info by ~20% (2.2MB) on
linux-image-4.4.0-trunk-rt-686-pae_4.4-1~exp1_i386.deb.

Note that objdump is still the primary memory hog for said package at
8.29MB (out of a total of 12.69).

Signed-off-by: Niels Thykier <niels@thykier.net>