Remove "." from @INC
It's a long-standing Perl bug (#588017) that "." is appended to @INC.
If a program tries to import module that is not installed system-wide,
Perl will attempt to import it from the current working directory. If
the directory is untrusted, this could lead to arbitrary code execution.

Lintian is affected, because it imports the Encode module, which tries
to import Encode::ConfigLocal, which doesn't normally exist.

To mitigate this bug, remove "." from @INC early from all Perl scripts
that can be (indirectly) run by end users.