Introduce a do_fork() sub to clear signal handlers
TL;DR: Without this Lintian turns into to a fork-bomb with an
ill-timed SIGINT/SIGTERM.

Lintian delegates all of the unpacking to the L::Unpacker module.
This module basically forks and lets the subprocess do an unpacking
job.  At a quick review of the code, it would seem that the subprocess
cannot "escape" from its scope and it would always either call
"POSIX::_exit" or "exec".

Nonetheless, there is a subtle race-condition post-forking if Lintian
would receive a SIGINT/SIGTERM.  The issue comes from the fact that
the subprocess inherits the signal handler of the parent process,
which uses a "print + die" to throw an exception[1].  Thus, in reality
*any* line in the child process can throw an exception (via a signal).

For an end user, this is generally not terribly visible.  When the
child escapes out, it will still cause the unpacking to fail.
Accordingly, the current package will not be checked and at worst the
clean up may disturb a currently running subprocess.

However, when lintian is processing multiple groups, the child now
escapes and starts the next group.  Which means we now have at least
*two* processes trying to do unpacking.  This is an issue that is
further agitated by reporting-lintian-harness now sending SIGTERM to
lintian, when it has not terminated within its 4 hour deadline.

This patch aims to solve all of that by introducing a new "do_fork"
sub to replace all uses of "fork()".  This sub will mask all signals,
fork, drop the signal handlers in the child and then restore the
original signalmask.

[1] This can be perceived locally by running lintian and trying to
interrupt it during unpacking (use --debug to see when it starts).  A
successful trigger will cause lintian to print *multiple*
"N: Interrupted" lines.  The excess lines comes from child
processes.

Signed-off-by: Niels Thykier <niels@thykier.net>