Give error for packages shipping php scripts but not depending on php-cli
The PHP maintainers recommend to have PHP dependencies without
specifying the PHP version explicitly, e.g.:

  Depends: php-cli, php-curl, php-xsl

instead of:

  Depends: php7.0-cli, php7.0-curl, php7.0-xsl

This is a safe thing to do considering that each Debian stable version
is going to ship with only one major version of PHP.

However in the case of packages shipping php scripts, lintian now
wrongly suggests to depend on a versioned phpX-cli package.

Fix this by triggering the lintian error when the package ships a script
and but does not depend on the unversioned php-cli dependency.

This also covers the deprecated case of using a versioned phpX-cli
dependency. To confirm that, a dependency to php7.0-cli has been added
to t/tests/legacy-scripts/debian/debian/control and it has been verified
that the package still gives the error.

Moving php from the "versioned interpreters" to the unversioned
interpreters also results in an "unusual-interpreter" warning when, for
example, an interpreter with a version number in the filename is used.

In a future commit the "unusual-interpreter" warning will be turned into
a more specific warning suggesting to use the unversioned php
interpreter in the shebang line.

The changes pass these tests:
  debian/rules runtests onlyrun=scripts-missing-dep
  debian/rules runtests onlyrun=legacy-scripts

Closes: #818962
Thanks: Mathieu Parent <Mathieu.PARENT@nantesmetropole.fr>

Signed-off-by: Antonio Ospite <ao2@ao2.it>
Signed-off-by: Niels Thykier <niels@thykier.net>