data/scripts/interpreter: fix false positive with python2 as interpreter
When using dh_python2 the package ends up depending on python:any
(Debian does not have a python2:any package).

However if a script uses python2 as the interpreter, the lintian
dependency checks fails:

  E: scripts: python-script-but-no-python-dep usr/bin/script_name

Basically lintian tries to look for a dependency on python2:any, it does
not know that python:any is OK as a dependency for scripts using python2
in the shebang line.

This can be verified by temporarily adding "python:any" to
t/tests/legacy-scripts/debian/debian/control and running the test:

  debian/rules runtests onlyrun=legacy-scripts

Lintian will not give the error anymore for pyfoo, but it will still
emit it for py2foo which is wrong.

Fix the issue adding python2 to the unversioned interpreters, the
rationale being that the python2 interpreter requires unversioned
dependencies.

After the fix, lintian will give a clearer error when the dependency is
really missing:

  E: scripts: missing-dep-for-interpreter python2 => python:any | python-minimal:any (usr/bin/py2foo)

And to prove that the false positive is not occurring anymore,
temporarily adding "python:any" to
t/tests/legacy-scripts/debian/debian/control will not give the error
anymore for py2foo either when running:

  debian/rules runtests onlyrun=legacy-scripts

Signed-off-by: Niels Thykier <niels@thykier.net>