t/runtests: Re-sort after running post_test script
This should fix a FTBFS on arm* and many other architectures without
having to disable the test "files-old-config-scripts".

Some outputs are architecture sensitive.  Most of the time, a simple
"sed s///" is sufficient to solve the problem.  However, in the
"files-old-config-scripts" test, the output for arm is sorted
differently than the output for amd64/i386 (relative to other output)
prior to the sed script is run.

To solve this, we could either run the sed script before sorting
(which would be "cheaper" resource-wise) or just re-sort post running
the script.  I have opt'ed for the latter on account of it being
easier to implement.

Signed-off-by: Niels Thykier <niels@thykier.net>