Drop debian-readme coll
It depends on unpacked, so it is not faster than unpacked.  It is
auto-removed (and even if it wasn't, we use a temporarily lab on
lintian.d.o), so the files are not retained.  Finally, the extracted
data is only used by one single check that can trivially implement
this logic itself.

This saves a stat calls, avoids some "low-level" code (which is more
likely to have information leak bugs) and finally it means there is
one less short-lived coll to load and run.

Signed-off-by: Niels Thykier <niels@thykier.net>