cmd/lintian: Avoid die() and reduce the (mis)usage of "fail"
Most of the error messages from cmd/lintian are entirely expected and
caused by user errors (e.g. invalid options).  Those errors should not
be classified as "uncaught exceptions" (die) or "internal errors"
(fail).

This commit corrects this by replacing most of uses of die/fail in
cmd/lintian with a new "fatal_error" that simply prints the message to
STDERR and exits with error code 2.

Signed-off-by: Niels Thykier <niels@thykier.net>