private/refresh-manual-refs: Don't trump earlier refs on section reuse
Some manuals reuse section numbers for different references, e.g. the
Debian Policy's normal and appendix sections are numbers that clash with
each other as of 4.1.1.1 (and seems that is not easily fixable in it).
Track if we've already seen a section pointing to some other URL than
the current one, and uniquify it by appending as many asterisks as
necessary.

Signed-off-by: Chris Lamb <lamby@debian.org>