d…/s…/corrections: Remove seldom, but valid English word "publically"
According to https://en.wiktionary.org/wiki/publically, "publically"
is a "(rare) Alternative spelling of publicly" which is indeed said to
be "generally considered more correct". But that doesn't make
"publically" incorrect.

Additionally "publically" is used in the last paragraph of the OpenSSL
license (c.f. https://spdx.org/licenses/OpenSSL and
https://www.openssl.org/source/license.html) and hence causes quite
some false positives for spelling-error-in-copyright. As of today,
https://lintian.debian.org/tags/spelling-error-in-copyright.html
showed 269 cases of which probably many are due to the OpenSSL
License.

Nevertheless, alone to avoid lintian overrides for hundreds of
packages being licensed under the OpenSSL license, this overzealous
spelling correction needs to be removed.

Additionally, its future re-addition should be avoided, too, for the
same reasons.