c…/scripts.desc: Improve desc. for command-with-path-in-maintainer-script
Bascially there are two cases of "if [ -x /usr/bin/<command> ]; …"
in maintainer scripts:

A) The maintainer controls the checked path (and doesn't forget to
   update the maintainer script in case he ever changes the program's
   path).

B) The maintainer doesn't control the checked path. (See #769845 and
   https://lists.debian.org/debian-devel/2014/11/msg00044.html for real
   examples.)

Overriding this tag in case (A) is ok if the maintainer indeed doesn't
forget to also update the maintainer scripts if he changes the path.

But case (B) is a real issue and should be fixed and not overridden,
because the maintainer script definitely behaves worse if the wrong
path (e.g. a path only valid in earlier Debian releases) is used in
the check.