Check for packages installing APT keyrings.
Add a new test "package-installs-apt-keyring", which checks for a
package installing /etc/apt/trusted.gpg or an /etc/apt/trusted.gpg.d/
file, with an exception for packages whose names end in -apt-source or
-archive-keyring. This ensures that a random package doesn't extend the
set of trusted keyrings.

Based on the current Contents file, this test will not warn on any
package currently in the Debian archive.

Signed-off-by: Chris Lamb <lamby@debian.org>