t/tests/debhelper-dh-nativejava: Remove obsolete test (-gcj packages are no longer built).
The test t/tests/debhelper-dh-nativejava test the functionality of debhelper
script dh_nativejava. At one point, the script was shipped in gcj-native-helper,
which is still listed as a test dependency. For a while, it was shipped in
libgcj-common. Then it was removed from Debian. According to N. Thykier,
the script is now obsolete since we no longer build -gcj packages.

Signed-off-by: Chris Lamb <lamby@debian.org>