checks/fields.pm: Fix a false-negative in the multi-arch-same-package-has-arch-specific-overrides tag. (Closes: #911355)
The Lintian test suite identified a false negative for the
'multi-arch-same-package-has-arch-specific-overrides' tag when the value of the
Architecture: field in 'debian/control' does not match the extension of the
lintian-override file. Both fields could be architecture wildcards. A positive
finding should take place when the latter is a subset of the former.

To reproduce, manually change both package architectures in the test
'fields-multi-arch-same-package-has-arch-specific-overrides' to Architecture:
any (in the file 'debian/debian/control.in'). The tag
'multi-arch-same-package-has-arch-specific-overrides' should disappear.

Signed-off-by: Chris Lamb <lamby@debian.org>