t/templates: Move old template changelogs out of the way. (MR: !57)
The template sets for Debian packages come with changelogs stored at
`debian/changelog`. They are not part of the template set but describe changes
to the templates. They reside just next to files like `debian/changelog.in`,
which are valid templates. The awkward location necessitates some
accommodations when copying templates. This patch moves the changelogs out of
the way.

Signed-off-by: Chris Lamb <lamby@debian.org>