Explain flexible templating feature in README (also includes two features already committed)
These edits to the README explain the general use and mechanism of the flexible
templating system.

This commit also includes information on the discontinued substitution of TESTSET in
Lintian options and how to declare prequisites for a test by using Extra-Build-Depends:,
which adds the dependency to Build-Depends, or Test-Depends:, which does not. Both
features were committed previously.

Signed-off-by: Chris Lamb <lamby@debian.org>