Use parameterized builder definitions according to suite when running tests.
By reading a special file called 'builder' the test runner determines the build
command and the build product. The runner prefills this special file before any
other templates and reads it into the data structure used for the remaining
templates.

This commit is another step in decoupling the build process from each suite. It
will ultimately allow the unification of the three separate test runners.

Signed-off-by: Chris Lamb <lamby@debian.org>