Use the make-based builder for suite tests, which was previously coded in Perl.
This commit causes the runner to use templates and skeletons updated for
make-based builders for the suite 'tests'. It will allow the unification of the
three test runners.

Extensive testing has not shown a speed disadvantage despite the overhead of
using 'make'. Should the need arise any other executable, including a compiled
binary, could serve as a builder. It would be configured solely via the
templating system, which can also be overridden by each test.

Signed-off-by: Chris Lamb <lamby@debian.org>