Do not use a global build stamp for tests in debian/rules. (MR: !80)
Using its Makefile semantics, debian/rules implements a build stamp
when tests are run without the option 'onlyrun'. It prevents needless
re-runs, but also blocks runs that should be allowed, for example when
only the test runner has changed. (The target 'runtests' depends on
many files, but not on the t/runtest executable or any of its
imports.) The mechanism is a complete obstacle to runner development.

This commit disables the mechanism.  The build stamp is no longer
written, and the 'runtest' target is declared PHONY.

Signed-off-by: Chris Lamb <lamby@debian.org>