In tests, allow /debian/rules to be generated from templates. (!MR: 81)
The whitelist for Debian packaging templates used in tests does not
include debian/rules.  In the test suite, these files occasionally
hardcode the name of the test.  This commit makes it possible to
provide instead debian/rules.in that is filled with the correct value
via a declaration like "pkg={$source}".

Signed-off-by: Chris Lamb <lamby@debian.org>