Remove wrong line number for fields-bugs-field-does-not-refer-to-debian-infrastructure.
The tag fields-bugs-field-does-not-refer-to-debian-infrastructure
includes a line number as a hint, but that number is incorrect.  In
the test fields-bugs-field-does-not-refer-to-debian-infrastructure,
the offending field was located in line 9 of the generated control
file and not in line 11, as the tag said. The line number appeared
unrelated. It was removed.

At the time of writing, the tag did not appear in the wild. The error
could not be corroborated outside the Lintian test suite.

Signed-off-by: Chris Lamb <lamby@debian.org>