Fix test generic-dh-make-2005 to work without external templates.
test to remain undisturbed "as a test of Lintian's handling of old ...
template packages". This commit stops the use of standard templates,
which in the past provided missing files.

During the build process, dh_clean is called. It errors out when no
compat file is present. To get the test to run, a compat file is now
generated through the pre_build script. This is done dynamically so
that the set of static files is not modified, as requested by the
original author.

Some tags appeared when the use of default templates was discontinued.
Those tags were added.

Signed-off-by: Chris Lamb <lamby@debian.org>