Transfer file permissions from templates to generated files.
The templating system does not preserve file permissions, but scripts
must be executable to run. This commit applies the file permissions of
the template to any generated file. Going forward, scripts generated
from templates can also run.

This change is important for a new mechanism to generate debian/watch
and is also used to supply compat files to the tests
generic-dh-make-*, which were created without it. Modern Debian tools
require compat. The script was a way to add it without modifying the
debian files originally created by dh-make.

Signed-off-by: Chris Lamb <lamby@debian.org>