Remove dynamic generation of watch file from builder Makefile.
For a long time, the Lintian test suite automatically created watch
files for non-native packages. That worked well, except in some corner
cases. Some tests compensated by removing the file in the pre_build
script.

This commit is part of a series that changes that mechanism. The watch
file is now supplied by a template set and removed right after the
templates are copied into a test's work directory.  The processing
takes place earlier. It allows tests deal with watch files like they
do with any other file that is (or is not) supplied by template sets.

Signed-off-by: Chris Lamb <lamby@debian.org>